lsf-configurator/templates/python/composition/requirements.txt
parliament-functions==0.1.0
requests
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-exporter-otlp
redis
